description: provide convenience top-levels and macros.
These avoid users having to provide values for *all* arguments of the 'ur_robot' macro.

In many cases, users will want to keep the default files for the joint limits, physical and visual parameters, and only override the default kinematics (to use extracted calibration fi).
With the provided wrapper macros, this is possible, as variant-specific defaults are provided for all arguments, requiring only to override the required ones.

Users looking to include a UR into a larger scene or composite xacro macro should 'xacro:include' these '_macro.xacro' files.

The top-levels are only useful when loading a stand-alone UR in an otherwise empty scene. They do not allow access to any arguments and only use the defaults.